requirements.txt
certifi==2018.10.15
cloudinary==1.14.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.4.0
django-hitcount==1.3.0
django-mathfilters==0.4.0
gunicorn==19.9.0
joblib==0.13.2
mock==2.0.0
numpy==1.16.4
pandas==0.23.4
pbr==5.1.0
Pillow==5.3.0
psycopg2==2.7.5
python-dateutil==2.7.5
pytz==2018.5
scikit-surprise==1.0.6
scipy==1.3.0
six==1.11.0
sklearn==0.0
surprise==0.1
urllib3==1.24
whitenoise==4.1
